moved group_live to separate folder
