+++ lib/haj_web/templates/group/_side_bar.html.heex
@@ -1,22 +1,22 @@
 <div class="flex flex-col">
   <div class="bg-gray-100 px-3 py-2 font-bold uppercase"><%= @show_group.group.name %></div>
   <div class="border-b border-gray-300 px-3 py-2">
-    <%= link("Medlemmar", to: Routes.group_path(@conn, :group, @show_group.id)) %>
+    <%= link("Medlemmar", to: Routes.group_index_path(@conn, :group, @show_group.id)) %>
   </div>
 
   <%= if is_admin?(@show_group, @conn.assigns.current_user) do %>
     <div class="border-b border-gray-300 px-3 py-2">
-      <%= link("Administrera", to: Routes.group_path(@conn, :edit, @show_group.id)) %>
+      <%= link("Administrera", to: Routes.group_index_path(@conn, :edit, @show_group.id)) %>
     </div>
     <div class="border-b border-gray-300 px-3 py-2">
-      <%= link("Ansökningar", to: Routes.group_path(@conn, :applications, @show_group.id)) %>
+      <%= link("Ansökningar", to: Routes.group_index_path(@conn, :applications, @show_group.id)) %>
     </div>
     <div class="border-b border-gray-300 px-3 py-2">
-      <%= link("Ladda ner csv", to: Routes.group_path(@conn, :csv, @show_group.id)) %>
+      <%= link("Ladda ner csv", to: Routes.group_index_path(@conn, :csv, @show_group.id)) %>
     </div>
   <% end %>
 
   <div class="border-b border-gray-300 px-3 py-2">
-    <%= link("Ladda ner vcard", to: Routes.group_path(@conn, :vcard, @show_group.id)) %>
+    <%= link("Ladda ner vcard", to: ~p"/show-groups/#{@show_group}/vcard") %>
   </div>
 </div>
