+++ lib/haj_web/templates/settings_html/show_groups/edit.html.heex
@@ -3,6 +3,6 @@
     <._side_bar current_spex={@current_spex} conn={@conn} />
   </div>
   <div class="flex-grow-[3]">
-    <%= live_render(@conn, HajWeb.GroupAdminLive, session: %{"show_group_id" => @show_group.id}) %>
+    <%= live_render(@conn, HajWeb.GroupLive.Admin, session: %{"show_group_id" => @show_group.id}) %>
   </div>
 </div>
